Software/web-server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
msgpack==1.0.7
stomp.py==8.1.0
pillow==10.1.0
pyyaml==6.0.1
aiofiles==23.2.1
websockets==12.0